+++ src/main/java/com/dyl/sell/dto/ClientWarehouseDetailed.java
@@ -61,7 +61,7 @@ public class ClientWarehouseDetailed {
     /**
      * 商品数量
      */
-    private Integer amount;
+    private Double amount;
     /**
      * 商品进价
      */
