+++ src/main/java/com/dyl/sell/repository/SellDescRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface SellDescRepository extends JpaRepository<SellDesc, String> {
-    public List<SellDesc> findAllOrderByAmount(Integer amount);
+    public List<SellDesc> findAllOrderByAmount(Double amount);
     public List<SellDesc> findAllOrOrderByTotalPrice(Double totalPrice);
 }
